refactor: Drop separate server.js file in favor of using webpack-dev-server directly through webpack config
